refactor: move navigation drawer to external component
