+++ frontend/components/Layout/Navigation/NavigationDrawer.vue
@@ -1,0 +1,102 @@
+<template>
+  <v-navigation-drawer
+    v-if="navDrawer"
+    v-model="drawer"
+    :temporary="$vuetify.breakpoint.mobile"
+    :permanent="!$vuetify.breakpoint.mobile"
+    app
+    class="pa-s"
+  >
+    <template #prepend>
+      <user-button />
+      <v-divider />
+    </template>
+    <v-list>
+      <v-list-item
+        v-for="item in items"
+        :key="item.Id"
+        :to="item.to"
+        router
+        exact
+      >
+        <v-list-item-action>
+          <v-icon>{{ item.icon }}</v-icon>
+        </v-list-item-action>
+        <v-list-item-content>
+          <v-list-item-title v-text="item.title" />
+        </v-list-item-content>
+      </v-list-item>
+      <v-subheader>{{ $t('libraries') }}</v-subheader>
+      <v-list-item
+        v-for="library in libraryItems"
+        :key="library.Id"
+        :to="library.to"
+        router
+        exact
+      >
+        <v-list-item-action>
+          <v-icon>{{ library.icon }}</v-icon>
+        </v-list-item-action>
+        <v-list-item-content>
+          <v-list-item-title v-text="library.title" />
+        </v-list-item-content>
+      </v-list-item>
+    </v-list>
+    <template #append>
+      <connection-monitor />
+      <syncing-monitor />
+      <v-list-item
+        v-if="commit"
+        :href="'https://github.com/jellyfin/jellyfin-vue/commit/' + commit"
+      >
+        <v-list-item-action>
+          <v-icon>mdi-github</v-icon>
+        </v-list-item-action>
+        <v-list-item-content>
+          <v-list-item-title v-text="'#' + commit.substring(0, 6)" />
+        </v-list-item-content>
+      </v-list-item>
+    </template>
+  </v-navigation-drawer>
+</template>
+
+<script lang="ts">
+import { BaseItemDto } from '@jellyfin/client-axios';
+import Vue from 'vue';
+import { mapState } from 'vuex';
+import { AppState } from '~/store';
+import { getLibraryIcon } from '~/utils/items';
+
+interface LayoutButton {
+  icon: string;
+  title: string;
+  to: string;
+}
+
+export default Vue.extend({
+  computed: {
+    ...mapState<AppState>({
+      libraryItems: (state: AppState) =>
+        state.userViews.views.map((view: BaseItemDto) => {
+          return {
+            icon: getLibraryIcon(view.CollectionType),
+            title: view.Name,
+            to: `/library/${view.Id}`
+          };
+        })
+    }),
+    commit() {
+      return process.env.NUXT_ENV_COMMIT;
+    },
+    items(): LayoutButton[] {
+      return [
+        {
+          icon: 'mdi-home',
+          title: this.$t('home'),
+          to: '/'
+        }
+      ];
+    }
+  }
+});
+</script>
